backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "poethepoet>=0.36.0",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
]


[tool.poe.tasks]
start = "fastapi dev main.py"
